Moved to utils and changed from Loding to Starting
